requirements.txt
flask==3.0.3
requests==2.32.3
PyPDF2==3.0.1
python-docx==1.1.2
gunicorn==22.0.0
